backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.10.0
pydantic-settings>=2.1.0
pydantic-ai>=0.0.14
openai>=1.12.0
httpx>=0.27.0
chromadb>=0.4.22
python-multipart>=0.0.6
python-dotenv>=1.0.0
numpy>=1.24.0
aiofiles>=23.0.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
pypdf>=3.17.0
python-docx>=1.1.0
